backend/requirements.txt
fastapi==0.95.0
uvicorn==0.18.3
pydantic==1.10.2
motor==2.5.0
python-dotenv==0.19.2
